Add files via upload
That project implement relation between controller and application for how mush controller should use a core. 
The application read a file and looking for amount of words in a file, and measure time that spend for that search. After that an application sent that information in a controller that will make a decision  about number of cores. If time less then delay that was given in beginning, we will not change an amount of core, otherwise we will increase. Maximum number of core is 4.